Fix image killed after deletion
If an image upload fails after the image was deleted during
the upload, the image status currently gets set to 'deleted'
then 'killed' once the upload fails. This patch fixes this so
that if the image was deleted it stays in 'deleted' state.

Permitted from_states for transitioning to 'killed' are
'queued' and 'saving'. We now enforce this transition and
raise an exception if it is violated.

Closes-Bug: 1236868
Co-authored-by: Hirofumi Ichihara <ichihara.hirofumi@lab.ntt.co.jp>
Co-authored-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>
(cherry picked from commit 345d8f7766aa06fa6654382efc480da493a3f2c8)

Conflicts:

	glance/api/v1/images.py
	glance/api/v1/upload_utils.py

Change-Id: I5552220ce264c0d09780bbd0eb7610e7cff0b820