Adding an image status transition diagram for dev doc
Now there is no image status transition diagram in the dev
doc, this patch will add one and keep the image source as
a dot file so that it can be updated easily in the future.
See http://www.graphviz.org to get more info about .dot file.

Fixes bug 1244995

Change-Id: I076c461f688be93eab8d3d1fa541990d7b79da5e