Add config option to limit image properties
This patch adds the image_property_quota config option. This allows a deployer
to limit the number of image properties allowed on an image. The default value
is 128, as is currently the limit enforced by nova. Users will only be able to
update an image if the result of the transaction would be under this limit.
This behavior is intended to be similar to 'quota_metadata_items' in nova.

This is for both Glance v1 and v2.

Fixes bug 1251518
docImpact

Change-Id: I4aa9504deae836404f11c9ada71a91f85caeba4c