Add config option to limit image tags
This patch adds the image_tag_quota config option. This allows a deployer
to limit the number of image tags allowed on an image for v2. The default value
is 128. If an image is somehow over the limit, tags can still be removed but no
more may be added.

Fixes bug 1252337
docImpact

Change-Id: I2036e2a88601d7c5aa85fad32d90fe0ed30b84c8