Replace startswith by more precise store matching
Currently a source is considered valid if it only starts with a specific set
of word. For example swift is a valid location scheme, so is swiftpe or
swiftopenstack with the current implementation. This commit compare the scheme
of the location - all of it - to a given set of supported schemes

Change-Id: Ie827c3c3f36fbf958a1270039cab28d4c5fc4fe1
Closes-Bug: #1257496