Disallow negative image sizes
Currently Glance v1 allows creation of images with negative sizes when
the source URL provided by the --location parmeter does not exist at
the time of creation.

There is already a check to verify whether or not an image size can be
converted to an int but the check doesn't verify whether or not the
resulting value is >= 0. This commit supplements that check to
disallow negative values for the size parameter.

Change-Id: I47511f3e34cf5adfdf944747d94285ecbbda4249
Closes-Bug: #1260738