Add config option to limit image locations
This patch adds the image_location_quota config option. This allows
a deployer to limit the number of image locations allowed on an image.
The default value is 10. Users will only be able to update an image
if the result of the transaction would be under this limit. This
behavior is intended to be similar to 'quota_metadata_items' in Nova.

We have implemented the same limit for image properties, members and tags,
see:
https://review.openstack.org/#/c/56981/
https://review.openstack.org/#/c/59563/
https://review.openstack.org/#/c/58056/

docImpact
Closes-Bug: #1259031

Change-Id: I0e0033a1d11be267144e8e10b730bb6331263d4c