Makes 'expires_at' not appear if not set on task
Fixes bug: 1267148

'expires_at' on a task must be set only after the executor succeeds or
fails on a task. Before the task has finished, there shouldn't be any
'expires_at' set on the task. This change fixes the ResponseSerializer
to not show 'expires_at' if it is not set.

Change-Id: I9880cc03b9d6f0dd98388fcf0d1bdc9c34dd6c0e