Move scrubber outside the store package
Glance's scrubber was created under the store package because it is
intended to clean up images from stores when `delayed_delete` is
enabled. However, it doesn't really belong there.

Soft delete is a Glance feature that allows images to be deleted
asynchronously as opposed of blocking clients requests during the image
deletion. Therefore, it needs to be store agnostic and live within
Glance's codebase.

Partially-Implements blueprint: create-store-package

Change-Id: I25ddbe1f0f1191f33d814dbae5788da7d511773a