Provide an upgrade period for enabling stores
1f6381a73f5c99f1f731d6c4f9defb91bd2d042d disabled all stores by default,
this casues upgrade issues for CD environments and folks that are not
aware of the change. This patch adds a way to attempt enabling the store
if its configuration parameters were set. If the store fails to load and
it's not in `known_stores` then no error will be reported.

The patch also warns the user when a store succeeds to load but it was
not explicitly enabled in `known_stores`. Note that there may be false
positives since some stores may not require any configuration parameter.

Change-Id: Ie784fe90003a5064aaf16fadb7266bf7a46e2cf5
Closes-Bug: #1290969