VMware store.add to return the image size uploaded
This patch fixes a tempest test failing for the VMware datastore.
When the size provided to add() is zero but the image size is actually
larger than zero, add() should return the size of the image uploaded
and not zero.

Tempest failure:
api.image.v1.test_images.CreateRegisterImagesTest.test_register_then_upload

Change-Id: I16524c0f25aa4124c24ca25ac014819267dd72c3
Closes-Bug: #1292764