make uploading an image as public admin only by default
Uploading an image with --is-public=True should by default only be allowed
by an admin tenant. Allowing anyone to upload an image as is_public is likely
a security concern. This changes the previous default behavior glance had
of allowing anyone to upload an image as is_public previously by default.

DocImpact
Closes-bug: 1317314

Change-Id: I60d2257115e9207e09b50d9f950076b7fe8237d7