Add pluging sample for glance gate
Recently we add support for plugins in gates, so now you are able
to keep rally benchmarks in your source tree. It may be usefull in cases
when you would like to check/debug in gates with benchmark that doens't exist
in Rally but don't want to spend whole life to get in Rally source.

As well we add extra/ directory that is just copy pasted in gates.
So we can use absolute path in task configs, cause this directory
will be copy pasted to ~/.rally/extra. This may be super usefull if you
would like to test infrastracture on scale, creating images from empty files.

Change-Id: I759ee8bf7a930169afd6076ec96f8fe3a3692866