Remove eventlet_hub option
The option has a default value that is not portable, and therefore the
test fails on OS like Darwin. There's actually no need to specify any
hub – no other OpenStack projects as that option – as in theory Eventlet
is smart enough to pick the best implementation available anyway.

However, since Glance only works with poll or select as a hub, let's try
them in that order. That avoids a user selecting things like 'epolls'
and expecting Glance to work, whereas it would fail.

Also remove a useless cleanup() call in the failing test as the
start_servers() method already does a cleanup itself.

Change-Id: Ia7765151ebb00cdf01e96cf39f3242899d358772