Add task functions to v2 registry
If local changes are made to run the v2 functional
tests with the v2 registry enabled, and the tests
are then run with:

$ ./run_tests.sh glance.tests.functional.v2

all tests pass except for 'test_task_lifecycle'.

This test fails because the v2 registry does not define the
'task_get_all' or 'task_create' functions.

With these defined the tests pass when run with the v2 registry enabled.

This is a prerequisite to running the v2 functional tests with the
registry enabled.

Change-Id: I588af10105b19087d06f7f13a6f75523595d4a23
Closes-Bug: 1308419