Prevent setting swift+config locations
Forbid setting 'swift+config' locations in a similar
manner to 'file' for security reasons; knowledge of
the reference name should not be exploitable.

Setting swift+config had been prevented when swift
was the default store, this patch changes to forbid
setting no matter which store is the default.

As with change id I75af34145521f533dcd6f5fd7690f5a68f3b44b3
this is v1 only for now.

Change-Id: I62c4980bd5c2f3dd77fc40cd007bc1067eca63a4
Closes-bug: 1334196