Log upload failed exception trace rather than debug
We're seeing race failures in test_register_http_image
in the gate due to a 500 coming back from the backing
store when uploading the image in the v1 API, but it's
logged at debug level so we can't fingerprint on the
failure in logstash.

This changes bumps the logging of the error from debug
to exception so we get the trace on the error also.

If it turns out that we have expected exception cases
in the future and tracing on those isn't necessary,
we can filter them out in a separate change.

Change-Id: I87d1f8ae584b41456eae4f19a0bc60f13f47b5ac
Related-Bug: #1347692