Check on schemes not stores
Since the `_upload` method loads the store based on the scheme, we need
to check whether the scheme exists in the known schemes instead of
checking in the list of known stores. Without this fix, stores with a
scheme name different than the store itself will not be usable through
the `x-image-meta-store` header.

Change-Id: Ia33afc1629cb47a31982df7932ac0d9aeb201b4f
Closes-bug: #1357462