Check on schemes not stores
Since the `_upload` method loads the store based on the scheme, we need
to check whether the scheme exists in the known schemes instead of
checking in the list of known stores. Without this fix, stores with a
scheme name different than the store itself will not be usable through
the `x-image-meta-store` header.

Conflicts:
	glance/api/v1/images.py

Closes-bug: #1357462
(cherry picked from commit 4502759243c0f363545f4236b41abaf8bcf9a37b)
Change-Id: Ia33afc1629cb47a31982df7932ac0d9aeb201b4f