remove url-path from the default value of s3_store_host
For s3 endpoints, url-path is not necessary to define and it's
something like just "s3.amazonaws.com" for Amazon s3 or
"s3.compatibleserver.com:8080" for S3 thirdparty compatible server
that is listening on 8080. So the current default value with '/v1.0/'
of s3_store_host is confusing.

s3_store_host = 127.0.0.1:8080/v1.0/

DocImpact: The following parameter is changed in glance-api.conf
and glance-cache.conf.
-s3_store_host = 127.0.0.1:8080/v1.0/
+s3_store_host = 127.0.0.1:8080

Change-Id: I3b103121e46f58fda339530c62c71b65a69e5dc5