Improve error log for expired image location url
If image is not present at the specified location while creating instance
from image, then HTTPInternalServerError 500 response along with stack trace
is logged on nova compute which does not help user to understand the exact
cause of failure.

Return HTTPNotFound error to the nova compute in case of image url got
expired or image is not present at the given location to give clear
indication of the cause of failure to user.

Closes-Bug: #1198566
Change-Id: I9acd9112aeae8d3b3c0c3921f306e716e5808c2e
(cherry picked from commit 633bec8fd45897735e1fcb9844903fe597903b21)