Use unicode for error message
Prior to this change, glance-reg would encounter an error, then
encounter an error reporting the first error. Glance-reg would end up
dead with the only error message printed being about the fact that the
first error message wasn't unicode.

After this change, glance-reg still dies just as readily - but on the
way down, it manages to emit the original error message. In my case,
this means that I now know that the original fault is that it can't bind
to 0.0.0.0:9191

Change-Id: I0e90b538996f87fe100c973fdf1fb61791e4e660
closes-bug: 1383255