Expose nullable fields properties
Some fields are not required parameters, therfore this patch allows them
to be None.

This has been done for images in a previous patch. This patch does the
same thing for tasks.

This will allow the client to have attributes with value `None` and
it'll make it consistent across responses and with regards to the API.

NOTE: A follow-up patch will enable None values to be returned by the
API.

Change-Id: I694cc144375c76ad8aaab495fc2e3ac5ed84ec2f