Cleanup chunks for deleted image that was 'saving'
Currently image data cannot be removed synchronously for an image that
is in saving state. And when, the upload operation for such an image is
completed the operator configured quota can be exceeded.
This patch fixes the issue of left over chunks for an image which was
deleted from saving status. However, by the limitation of the design we
cannot enforce a global quota check for the image in saving status.
This change introduces a inconsonance between http response codes of
v1 and v2 APIs. The status codes which we will now see after the upload
process completes on an image which was deleted mid way are:

v1: 412 Precondition Failed
v2: 410 Gone

SecurityImpact
UpgradeImpact
APIImpact

Closes-Bug: 1383973
Closes-Bug: 1398830
Closes-Bug: 1188532

Conflicts:
        glance/api/v1/upload_utils.py
        glance/api/v2/image_data.py
        glance/quota/__init__.py
        glance/tests/unit/test_domain_proxy.py
        glance/tests/unit/test_quota.py
        glance/tests/unit/v1/test_api.py

Signed-off-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>
(cherry picked from commit 0dc8fbb3479a53c5bba8475d14f4c7206904c5ea)

Change-Id: I47229b366c25367ec1bd48aec684e0880f3dfe60