Replace snet config with endpoint config
The snet option forces the deployer to name the desired endpoint after
the public endpoint. In order to switch between multiple internal
networks, names have to be changed. Instead of constructing a URL with a
prefix from what is returned by auth, specify the URL via configuration.

DocImpact
UpgradeImpact

bp replace-snet-config-with-endpoint-config

Change-Id: I7b7e14d7da082ae37a9737d2078017b04c19b87b