Zero downtime configuration reload
Add the ability to dynamically change configuration settings of a running
glance server with no impact to service.

A running glance server consists of a parent/master process and one or
more child/worker processes.

On receipt of a SIGHUP signal the master process will:

* reload the configuration
* send a SIGHUP to the original workers
* start (a potentially different number of) new workers with the new
  configuration
* its listening socket will *not* be closed

On receipt of a SIGHUP signal each original worker process will:

* close the listening socket so as not to accept new requests
* complete any in-flight requests
* complete async requests (V1 create with copy-from option and V2 task api)
* exit

This approach is based on nginx's behaviour and avoids some of the
disadvantages of the current oslo Launcher reload:

* Race conditions: Launcher does not shutdown eventlet cleanly, existing
  requests can fail.
* If all workers are busy (likely if #workers is small) there can be a
  lengthy delay when new requests are not processed.
* Long lived pre-SIGHUP idle client connections can stall request
  processing indefinitely.
* Not all parameters can be changed, eg number of workers.
* The wsgi pipeline cannot be changed, eg to enable caching.

Change-Id: I73875c654410f207e3efdb2932147cf26d75b3f1
Partially implements: bp sighup-conf-reload