Add detail description of image_cache_max_size
The description of 'image_cache_max_size' was ambiguous. There was not
the certainty that the 'image_cache_max_size' parameter was about the
aggregate size of the cache or a per image limit.
The 'image_cache_max_size' is an upper limit beyond which pruner,
if running, starts cleaning the images cache.
Hence added detail description of 'image_cache_max_size' in document
and configuration file.

Bug: #1411813
Change-Id: Ide8e087db544aeea1990bac92c97cb9ca9f0b522