Synchronising oslo-incubator service module
Importing service module and its dependencies to enable
openstack common service launcher within glance. This will
be used as the framework for event notifiers and listeners.

This change is a pre-requisite for blueprint
catalog-index-service implementation

All imported modules are new and shouldn't impact any of
the existing glance functionality.

Closes-Bug: #1413861
Partially Implements: blueprint catalog-index-service

Change-Id: Iaa771ead0114e3941667b1e07ff32472d2f77afd