Image data remains in backend for deleted image
Trying to delete image created using task api (import-from) image gets
deleted from the database, but image data remains in the backend. Import
task does not update the location of the image and it remains None even
image becomes active. Location entry is not added in the database in
image_locations table.

Added location information to the image before saving the image in
the database.

SecurityImpact

Change-Id: Ie389de6538a9b98dc51c7d781b81b3ab10b83842
Closes-Bug: #1420696