Purge dead file-backed scrubber queue code
The file-backed scrubber can no longer be used (see 66d24bb), but the
code for it is still in place, including several configuration
options, abstraction layers, and so on. This commit removes that dead
code.

It also deprecates the scrubber_datadir, cleanup_scrubber, and
cleanup_scrubber_time options, which are no longer
functional. "Cleanup" mode is also removed, as it is now equivalent to
the normal running mode of the scrubber.

DocImpact
Closes-Bug: 1427929
Change-Id: I3c2a43acee21c685a2de3a7b67ba831548bc27df