Avoid usability regression when generating config
The descriptions of some of the configuration options had been updated
in the sample config, but not in the declaration of the oslo.config
option. This patch maintains parity between the two places these are
currently defined, so that when we switch to only generating the config
files, there isn't a loss of information.

Change-Id: I634c03ac89a0f78cf86992490eb44d2db7a90352