Move elasticsearch dep to test-requirements.txt
Move experimental feature deps to test-requirements Optional and experimental
features should not make glance dependant on the libraries they require. This
change moves the python bindings for elasticsearch from the main
requirements.txt to test-requirements.txt.

semantic_version should potentially also be moved here, but it is currently
required to do a db sync and is generally less alarming than an elasticsearch
dependency. It is instead moved to the bottom of requirements.txt and marked as
being artifacts related.

Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>

Change-Id: I7f8dfafdb11dbf7ba258e907fbd17d92319d0f52
Closes-bug: #1441239