Remove ordereddict from requirements
Ordereddict was required for py26 support. Since we have
dropped support for py26 (we are no longer gating on it) we
should remove it as a requirement and remove references
to it in the code.

Change-Id: I648edbb5d0b51c003f08f5f5c4ce700128032b38
Partial-bug: 1445827
(cherry picked from commit 20ec82f18c368bbc8a89e79d2d6eee2de2c2a9ea)