Make sure the converted image is imported
Turns out the converted image is being completely ignored. This is
because the `file_path` provided by the ImportToFS task cannot be
overriden by other tasks. In order to fix this, I've dropped all the
extra suffixes that we've been using in favor of working on the actual
`file_path` directly.

Change-Id: Ifbd2308057f33c1f712fa06f9bb3e57f90068fdc
Closes-bug: #1452750