Fix delayed activation without disk and containers formats
If we want to perform a delayed activation of an image without
providing disk format or container format, Glance tries to
upload a file and fails with 400 error.

This code handles this situation by checking that no sources
were provided, and returns information about the image before
starting the upload.

Change-Id: Id6f81df9bb43697b0d694528b010bb3c7c9bedb3
Closes-Bug: #1268680